requirements.txt
Flask==3.1.0
opencv-python==4.10.0.84
picamera2==0.3.24
RPi.GPIO==0.7.1
firebase-admin==6.5.0
face-recognition==1.3.0
